Some light refactor
